requirements.txt
psycopg2-binary
requests
tqdm
python-dateutil
ujson
python-dotenv